requirements.txt
Flask==0.10.1
setuptools==33.1.1
Flask-Script==2.0.5

Flask-Testing==0.4.2
ipdb==0.8
ipython==2.1.0
mock==1.0.1
nose==1.3.4
nose-exclude==0.4.1
nose-parameterized==0.5.0
coverage==3.7.1
python-dateutil=2.7.3
